+++ image.go
@@ -188,15 +188,3 @@ func drawBackground(img *image.RGBA, posX, posY int, label string, col color.RGB
 		}
 	}
 }
-
-// maxStringWidth は表示上のテキストの最も幅の長い長さを返却する。
-func maxStringWidth(s []string) (max int) {
-	for _, v := range s {
-		text := getText(v)
-		width := runewidth.StringWidth(text)
-		if max < width {
-			max = width
-		}
-	}
-	return
-}
